Split configurations for API/data
