+++ apps/argos_api/lib/argos_api/document_controller.ex
@@ -1,6 +1,6 @@
 defmodule ArgosAPI.DocumentController do
 
-  @elasticsearch_url Application.get_env(:argos_data, :elasticsearch_url)
+  @elasticsearch_url Application.get_env(:argos_api, :elasticsearch_url)
   import Plug.Conn
 
   def get(conn) do
