+++ apps/argos_data/lib/argos_data/release.ex
@@ -1,8 +1,4 @@
 defmodule ArgosData.Release do
-  require Logger
-  @elasticsearch_url Application.get_env(:argos_data, :elasticsearch_url)
-  @elasticsearch_mapping_path Application.get_env(:argos_data, :elasticsearch_mapping_path)
-
   def seed_projects(date) do
     HTTPoison.start()
     ArgosData.ProjectCLI.run(date)
@@ -11,27 +7,5 @@ defmodule ArgosData.Release do
   def seed_projects() do
     HTTPoison.start()
     ArgosData.ProjectCLI.run()
-  end
-
-  def update_mapping() do
-    HTTPoison.start()
-    mapping = File.read!(@elasticsearch_mapping_path)
-
-    clear_index()
-
-    "#{@elasticsearch_url}/_mapping"
-    |> HTTPoison.put(mapping, [{"Content-Type", "application/json"}])
-    |> IO.inspect
-  end
-
-
-  def clear_index() do
-    HTTPoison.start()
-
-    "#{@elasticsearch_url}"
-    |> HTTPoison.delete()
-
-    "#{@elasticsearch_url}"
-    |> HTTPoison.put()
   end
 end
