+++ config/dev.exs
@@ -1,7 +1,8 @@
+# This file is responsible for configuring your application
+# and its dependencies with the aid of the Mix.Config module.
+#
+# This configuration file is loaded before any dependency and
+# is restricted to this project.
+
+# General application configuration
 use Mix.Config
-
-config :argos_api,
-  elasticsearch_url: "localhost:9200/argos"
-
-  # projects_url: "ocalhost:4000",
-  # projects_harvest_interval: 1000 * 60 * 5 # 5 minutes
